LayoutTests/fast/forms/color/input-color-onchange-event-expected.txt

Test if change event fires properly when color chooser changes. Bug 66848
To manually test this, click on the input color element in the top left corner and change the value from the color chooser. See if the number of "value changed" messages matches the number of times you changed the color.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS onChangeCount is 0
input event dispatched - value is:  #ff0000
change event dispatched - value changed to #ff0000
PASS onChangeCount is 1
PASS onChangeCount is 1
PASS input.value is "#ff0000"
PASS successfullyParsed is true

TEST COMPLETE
